backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==0.21.1
requests==2.28.2
PyYAML==6.0.1
gunicorn==21.2.0
